0326/requirements.txt
langchain
langchain-community
faiss-cpu
python-dotenv
openai